Auth_service/app/requirements.txt
flask
flask-sqlalchemy
requests
pyjwt
pycryptodome
cryptography
